// backend/src/entities/prisma/exchanges.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum Exchanges {
    ALL = "ALL",
    BINANCE = "BINANCE",
    MEXC = "MEXC",
    OKX = "OKX",
    COINBASE = "COINBASE",
    COINBASEEXCHANGE = "COINBASEEXCHANGE",
    COINBASEINTERNATIONAL = "COINBASEINTERNATIONAL",
    KRAKEN = "KRAKEN",
    KRAKENFUTURES = "KRAKENFUTURES",
    BYBIT = "BYBIT",
    BITGET = "BITGET",
    GATE = "GATE",
    HUOBI = "HUOBI",
    HTX = "HTX",
    KUCOIN = "KUCOIN",
    KUCOINFUTURES = "KUCOINFUTURES",
    CRYPTOCOM = "CRYPTOCOM",
    BITFINEX = "BITFINEX",
    BITMEX = "BITMEX",
    BITSTAMP = "BITSTAMP",
    GEMINI = "GEMINI",
    BITMART = "BITMART",
    BITRUE = "BITRUE",
    ASCENDEX = "ASCENDEX",
    PROBIT = "PROBIT",
    POLONIEX = "POLONIEX",
    LBANK = "LBANK",
    PHEMEX = "PHEMEX",
    WOO = "WOO",
    WOOFIPRO = "WOOFIPRO",
    DERIBIT = "DERIBIT",
    BINGX = "BINGX",
    HASHKEY = "HASHKEY",
    COINEX = "COINEX",
    WHITEBIT = "WHITEBIT",
    XT = "XT",
    MEXC3 = "MEXC3",
    P2B = "P2B",
    TRADEOGRE = "TRADEOGRE",
    NDAX = "NDAX",
    OXFUN = "OXFUN",
    BLOFIN = "BLOFIN",
    COINCATCH = "COINCATCH",
    BINANCEUS = "BINANCEUS",
    BINANCEUSDM = "BINANCEUSDM",
    BINANCECOINM = "BINANCECOINM",
    OKCOIN = "OKCOIN",
    MYOKX = "MYOKX",
    OKXUS = "OKXUS",
    BITHUMB = "BITHUMB",
    UPBIT = "UPBIT",
    COINONE = "COINONE",
    HUOBIJP = "HUOBIJP",
    BITFLYER = "BITFLYER",
    COINCHECK = "COINCHECK",
    BITBANK = "BITBANK",
    ZAIF = "ZAIF",
    BTCBOX = "BTCBOX",
    INDODAX = "INDODAX",
    TOKOCRYPTO = "TOKOCRYPTO",
    COINSPH = "COINSPH",
    NOVADAX = "NOVADAX",
    MERCADO = "MERCADO",
    BITSO = "BITSO",
    BTCTURK = "BTCTURK",
    BTCALPHA = "BTCALPHA",
    EXMO = "EXMO",
    BITTEAM = "BITTEAM",
    KUNA = "KUNA",
    LATOKEN = "LATOKEN",
    HYPERLIQUID = "HYPERLIQUID",
    VERTEX = "VERTEX",
    PARADEX = "PARADEX",
    DERIVE = "DERIVE",
    APEX = "APEX",
    DEFX = "DEFX",
    WOOFIPRO_DEX = "WOOFIPRO_DEX",
    IDEX = "IDEX",
    WAVESEXCHANGE = "WAVESEXCHANGE",
    MODETRADE = "MODETRADE",
    ALPACA = "ALPACA",
    BEQUANT = "BEQUANT",
    BIGONE = "BIGONE",
    BIT2C = "BIT2C",
    BITBNS = "BITBNS",
    BITOPRO = "BITOPRO",
    BITVAVO = "BITVAVO",
    BL3P = "BL3P",
    BLOCKCHAINCOM = "BLOCKCHAINCOM",
    BTCMARKETS = "BTCMARKETS",
    CEX = "CEX",
    COINLIST = "COINLIST",
    COINMATE = "COINMATE",
    COINMETRO = "COINMETRO",
    COINSPOT = "COINSPOT",
    CRYPTOMUS = "CRYPTOMUS",
    DELTA = "DELTA",
    DIGIFINEX = "DIGIFINEX",
    ELLIPX = "ELLIPX",
    FMFWIO = "FMFWIO",
    HOLLAEX = "HOLLAEX",
    INDEPENDENTRESERVE = "INDEPENDENTRESERVE",
    LUNO = "LUNO",
    OCEANEX = "OCEANEX",
    ONETRADING = "ONETRADING",
    PAYMIUM = "PAYMIUM",
    TIMEX = "TIMEX",
    YOBIT = "YOBIT",
    ZONDA = "ZONDA",
    HITBTC = "HITBTC",
    HUOBI_LEGACY = "HUOBI_LEGACY"
}


registerEnumType(Exchanges, { name: 'Exchanges', description: undefined })